requirements.txt
flask==3.1.1
pika==1.3.2
python-dotenv==1.1.1
requests==2.32.3
ffmpeg-python==0.2.0
pyexiftool==0.5.6
yt-dlp
